Revert "hw_8_exceptions"
