+++ src/Main.java
@@ -1,48 +1,27 @@
-package main.java;
-
-import main.java.exceptions.OwnMoreThanHundredException;
-import main.java.exceptions.OwnNullObjectException;
-import main.java.exceptions.OwnOddException;
-import main.java.exceptions.OwnPalindromeException;
-
 import java.util.Scanner;
-import java.util.regex.Pattern;
 
 class Main {
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws Exception {
         Object obj1 = new Object();
         Object obj2 = null;
         try
                 (Scanner scanner = new Scanner(System.in)) {
             System.out.println("Enter your word: ");
-            String word = scanner.next(Pattern.compile("^[a-zA-Z]+$"));
-            try {
-                DataServer.checkPalindrome(word);
-            } catch (OwnPalindromeException e) {
-                e.printStackTrace();
-            }
+            String word = scanner.nextLine();
+            DataServer.checkPalindrome(word);
             System.out.println("<================================================================================================>");
-            scanner.nextLine();
             System.out.println("Enter your number: ");
             int number = scanner.nextInt();
-            try {
-                DataServer.checkNumber(number);
-            } catch (OwnMoreThanHundredException | OwnOddException e) {
-                e.printStackTrace();
-            }
+            DataServer.checkNumber(number);
             System.out.println("<================================================================================================>");
             scanner.nextLine();
-            System.out.println("You have a two objects: obj1 and obj 2, which one you want to check if it's a null (type 1 or 2)?");
+            System.out.println("You have a two objects: obj1 and obj 2, which one you want to check if it's null (type 1 or 2)?");
             int choice = scanner.nextInt();
-            try {
-                if (choice == 1) {
-                    DataServer.checkObject(obj1);
-                } else if (choice == 2) {
-                    DataServer.checkObject(obj2);
-                }
-            }catch (OwnNullObjectException e){
-                e.printStackTrace();
+            if (choice == 1) {
+                DataServer.checkObject(obj1);
+            } else if (choice == 2) {
+                DataServer.checkObject(obj2);
             }
             System.out.println("<================================================================================================>");
             scanner.nextLine();
