+++ src/exceptions/OwnMoreThanHundredException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class OwnMoreThanHundredException extends Exception{
+    public OwnMoreThanHundredException (String message){
+        super(message);
+    }
+}
