+++ src/exceptions/OwnOddException.java
@@ -1,4 +1,4 @@
-package main.java.exceptions;
+package exceptions;
 
 public class OwnOddException extends Exception{
     public OwnOddException(String message) {
