--- src/main/java/exceptions/OwnMoreThanHundredException.java
@@ -1,7 +1,0 @@
-package main.java.exceptions;
-
-public class OwnMoreThanHundredException extends Exception {
-    public OwnMoreThanHundredException(String message) {
-        super(message);
-    }
-}
